fefactor: frontend to feature-based directory structure
- reorganized components, hooks, and types into feature-based folders for circuit and gates
- moved shared UI and layout components to common and layout directories
- update imports throughout the codebase to reflect new paths
- remove obsolete index files and consolidated style/configuration constants into feature-specific files
