+++ frontend/src/components/common/ModeToggle.tsx
@@ -1,13 +1,13 @@
 import { Moon, Sun } from "lucide-react"
 
-import { Button } from "@/components/ui/button"
+import { Button } from "@/components/ui/button.tsx"
 import {
     DropdownMenu,
     DropdownMenuContent,
     DropdownMenuItem,
     DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
-import { useTheme } from "@/components/ThemeProvider"
+} from "@/components/ui/dropdown-menu.tsx"
+import { useTheme } from "@/providers/ThemeProvider.tsx"
 
 export function ModeToggle() {
     const { setTheme } = useTheme()
