+++ frontend/src/components/layout/SidePanelLayout.tsx
@@ -1,6 +1,6 @@
 import { Children, isValidElement, type ReactNode, useState} from 'react';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card.tsx';
+import { Button } from '@/components/ui/button.tsx';
 import { ChevronLeft, ChevronRight, type LucideIcon } from 'lucide-react';
 
 interface SidePanelLayoutProps {
