+++ frontend/src/features/circuit/constants.ts
@@ -1,0 +1,6 @@
+export const CIRCUIT_CONFIG = {
+    defaultNumQubits: 3,
+    defaultMaxDepth: 10,
+    qubitLabelWidth: 50,
+    footerHeight: 50
+};
