+++ frontend/src/features/gates/types.ts
@@ -1,0 +1,8 @@
+export interface Gate {
+    id: string;
+    name: string;
+    symbol: string;
+    color: string;
+    description: string;
+    qubits: number;
+}
