+++ frontend/src/features/inspector/components/QasmEditor.tsx
