+++ frontend/src/lib/api/client.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-export const axiosInstance = axios.create({
+export const api = axios.create({
     baseURL: import.meta.env.VITE_API_BASE_URL,
     headers: {
         'Content-Type': 'application/json',
