--- frontend/src/pages/Home/index.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/pages/Home/HomePage'
