--- frontend/src/pages/Workspace/index.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/pages/Workspace/WorkspacePage'
