--- frontend/src/types/circuit.ts
@@ -1,9 +1,0 @@
-import type { Gate } from "@/types/gates.ts";
-
-export interface DraggableGate {
-    id: string;
-    gate: Gate;
-    depth: number;
-    qubit: number;
-    isPreview?: boolean;
-}
